refactor: create and use custom hook useProductActions
